chore : move guard / strategy to common directory
