+++ src/common/strategy/jwt-refresh.strategy.ts
@@ -2,7 +2,7 @@ import {Injectable} from '@nestjs/common';
 import {PassportStrategy} from '@nestjs/passport';
 import {ExtractJwt, Strategy} from 'passport-jwt';
 import {ConfigService} from '@nestjs/config';
-import {UserService} from './../../user/user.service';
+import {UserService} from '../../user/user.service';
 
 @Injectable()
 export class JwtRefreshStrategy extends PassportStrategy(Strategy, 'jwt-refresh') {
