+++ src/common/strategy/local.strategy.ts
@@ -1,8 +1,8 @@
 import {Strategy} from 'passport-local';
 import {PassportStrategy} from '@nestjs/passport';
 import {BadRequestException, Injectable} from '@nestjs/common';
-import {AuthService} from '../auth.service';
-import {Err} from './../../common/error';
+import {AuthService} from '../../auth/auth.service';
+import {Err} from '../error';
 
 @Injectable()
 export class LocalStrategy extends PassportStrategy(Strategy, 'local') {
